[EL-395] change type of rating to double and set hieght of picture

+++ lib/ui/base/widget/slider_widget.dart
@@ -26,7 +26,9 @@ Widget SliderWidget(
 }
 
 Widget RateingBar(BuildContext context,
-    {required int starValue, required double barpercent, int rateTotal = 0}) {
+    {required double starValue,
+    required double barpercent,
+    int rateTotal = 0}) {
   return Row(
     children: [
       StarRate(context, starValue),
